# backend/chalicelib/public/swagger_model.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "backend"
    },
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/user/token": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "schema": {
                            "$ref": "#/definitions/Tokens"
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "summary": "Retrieve User Tokens from refresh token",
                "description": "Provide refresh-token in cookie and use to retrieve access tokens"
            }
        },
        "/user/info": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "antwon_user_pool": [
                            "email",
                            "openid",
                            "profile"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "schema": {
                            "$ref": "#/definitions/UserInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "summary": "Retrieve User Information about current user",
                "description": "Provides all user information, including name and email"
            }
        },
...
    },
    "definitions": {
        "Tokens": {
            "type": "object",
            "properties": {
                "id_token":  {"type":  "string"},
                "access_token":  {"type":  "string"},
                "token_type":  {"type":  "string"},
                "expires_in":  {"type":  "integer"}
            }
        },
        "UserInfo": {
            "type": "object",
            "properties": {
                "sub":  {"type":  "string", "example": "1deb5a59-435d-49b8-be68-daf5744587fa"},
                "email_verified":  {"type":  "boolean", "example": "true"},
                "name":  {"type":  "string", "example": "Albert"},
                "email":  {"type":  "boolean", "example": "ablert@antwon.dj"},
                "username":  {"type":  "string", "example":  "google_115636318877645090402"},
                "is_spotify_connected":  {"type":  "boolean", "example":  "true"},
                "room_code":  {"type":  "string", "example":  "SOIREE"}
            }
        },
...
    },
    "securityDefinitions": {
        "antwon_user_pool": {
            "in": "header",
            "type": "apiKey",
            "name": "Authorization",
            "x-amazon-apigateway-authtype": "cognito_user_pools",
            "x-amazon-apigateway-authorizer": {
                "type": "cognito_user_pools",
                "providerARNs": [
                    "arn:aws:cognito-idp:eu-west-2:303078101535:userpool/eu-west-2_Y4hA2uEzU"
                ]
            }
        }
    }
}